requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.17.0
SQLAlchemy==1.4.5
Werkzeug==1.0.1
WTForms==2.3.3
